Removing the labels from the generated PR.
As described in the workflow, the two types of available PATs do not
allow the ghcli to add tags (or modify the PR at all after filing).  The
new, fine grained tokens should, at least in theory. This appears to
still be broken after more than a year.  The classic PATs complain that
the user does not have permission to update the PR.  In testing this on
my account, the classic PAT works correctly, so I suspect what we're
seeing here is that the generated PR is in the Opencast (project)'s
namespace, where the user generating the PR is opencastproject - an
entirely separate entity than the organization.  Also, the organization
can't run workflows, so we can't do it that way either.

Fun times.