Translate strings in subtitle select
When creating a new subtitle via the dropdown, the
available languages would be displayed based on the
users browser language. This commit changes it so
that languages are displayed based on the language selected
by the user for the editor (which is still the browser language
in case the user did not select anything).