Show track selection for single videos
We are currently not showing the track selection in case
there is only one video. This made sense
when only a whole video
could be selected or deselected, but now the track selection
differentiates between the video and audio track of a video.
Therefore, we should also show the track selection for single videos.