Show error if no video tracks present
As of Opencast 18, Opencast officially supports events with audio-only
files, meaning events are not guaranteed to have at least one
track with video anymore. The latter
is however an implicit assumption in
our app and it will not work for audio-only
events. Until we fix this, we should at least display an error message
to users instead of letting them stumble around in a half
broken application.